VIP cycle
The view controller then passes the inputs to the interactor to get
some work done. Next, the interactor passes the output to the presenter
for formatting. Finally, the presenter asks the view controller to
display the results.

This is the flow of control and it is always in one direction. The VIP
cycle is not named VIP because of a very important person. It is
because there is an order to things. V then I then P.